refactor: remove IndividualMailboxSettings
